Review feedback: extract components from HomePage.tsx
