+++ src/main/reactjs/src/pages/HomePage.tsx
@@ -1,22 +1,9 @@
-import {
-  Box,
-  Grid,
-  TableCell,
-  Checkbox,
-  Paper,
-  TableHead,
-  TableRow,
-} from '@mui/material';
-import { FC, useState } from 'react';
-import { useTranslation } from 'react-i18next';
+import { Box, Grid, Paper } from '@mui/material';
+import { FC } from 'react';
 
-import {
-  StyledTableRow,
-  PaginatedTable,
-  Selectable,
-} from 'components/tables/Table';
-import { H1, H3, Text } from 'components/elements/Text';
+import { H1, Text } from 'components/elements/Text';
 import { TranslatorDetails } from 'interfaces/translator';
+import { TranslatorListing } from 'components/translators/TranslatorListing';
 
 const testData: Array<TranslatorDetails> = [
   {
@@ -45,102 +32,6 @@ const testData: Array<TranslatorDetails> = [
   },
 ];
 
-const ListingHeader = ({
-  selectedItems,
-  totalItems,
-  toggleAllSelected,
-}: {
-  selectedItems: number;
-  totalItems: number;
-  toggleAllSelected(selected: boolean): void;
-}) => {
-  const { t } = useTranslation();
-
-  return (
-    <TableHead>
-      <TableRow>
-        <TableCell padding="checkbox">
-          <Checkbox
-            checked={selectedItems == totalItems}
-            indeterminate={selectedItems > 0 && selectedItems < totalItems}
-            onChange={(event) => toggleAllSelected(event.target.checked)}
-          ></Checkbox>
-        </TableCell>
-        <TableCell>
-          <H3>{t('akt.translator.name')}</H3>
-        </TableCell>
-        <TableCell>
-          <H3>{t('akt.translator.languagePairs')}</H3>
-        </TableCell>
-        <TableCell>
-          <H3>{t('akt.translator.hometown')}</H3>
-        </TableCell>
-      </TableRow>
-    </TableHead>
-  );
-};
-
-const TranslatorListing = ({
-  translators,
-}: {
-  translators: Array<TranslatorDetails>;
-}) => {
-  const [selected, setSelected] = useState<Set<number>>(new Set());
-
-  const translatorDetailsRow = (
-    { name, languagePairs, hometown }: TranslatorDetails,
-    { selected, toggleSelected }: Selectable
-  ) => {
-    return (
-      <StyledTableRow selected={selected} onClick={toggleSelected}>
-        <TableCell padding="checkbox">
-          <Checkbox checked={selected} />
-        </TableCell>
-        <TableCell>
-          <Text>{name}</Text>
-        </TableCell>
-        <TableCell>
-          {languagePairs.map(({ from, to }, j) => (
-            <Text key={j}>
-              {from} - {to}
-            </Text>
-          ))}
-        </TableCell>
-        <TableCell>
-          <Text>{hometown}</Text>
-        </TableCell>
-      </StyledTableRow>
-    );
-  };
-
-  const toggleAllSelected = (allSelected: boolean) => {
-    if (allSelected) {
-      setSelected(new Set(Array.from(new Array(translators.length).keys())));
-    } else {
-      setSelected(new Set());
-    }
-  };
-  return (
-    <Paper elevation={3}>
-      <PaginatedTable
-        selectedIndices={selected}
-        setSelectedIndices={setSelected}
-        data={translators}
-        getRowDetails={translatorDetailsRow}
-        header={
-          <ListingHeader
-            selectedItems={selected.size}
-            totalItems={translators.length}
-            toggleAllSelected={toggleAllSelected}
-          />
-        }
-        initialRowsPerPage={10}
-        rowsPerPageOptions={[10, 20, 50]}
-      />
-    </Paper>
-  );
-};
-
 export const HomePage: FC = () => (
   <Box sx={{ width: 800, padding: '2em' }}>
     <Grid container rowSpacing={2} direction="column">
@@ -160,15 +51,17 @@ export const HomePage: FC = () => (
         </Box>
       </Grid>
       <Grid item>
-        <TranslatorListing
-          translators={[
-            ...testData,
-            ...testData,
-            ...testData,
-            ...testData,
-            ...testData,
-          ]}
-        />
+        <Paper elevation={3}>
+          <TranslatorListing
+            translators={[
+              ...testData,
+              ...testData,
+              ...testData,
+              ...testData,
+              ...testData,
+            ]}
+          />
+        </Paper>
       </Grid>
     </Grid>
   </Box>
